added user sepcific wo, route protection, middleware

+++ frend/src/components/WorkoutDetails.js
@@ -1,4 +1,5 @@
-import useWorkoutContext from '../hooks/useWorkoutContext'
+import { useWorkoutContext} from '../hooks/useWorkoutContext'
+import {useAuthContext} from '../hooks/useAuthContext'
 
 // date-fns
 import { formatDistanceToNow } from 'date-fns'
@@ -7,10 +8,19 @@ import { formatDistanceToNow } from 'date-fns'
 const WorkoutDetails = ({ workout }) => {
 
   const { dispatch } = useWorkoutContext()
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+
+    if (!user) {
+      return
+    }
+
     const response = await fetch(`/api/workouts/${workout._id}`, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+    }
     })
     const json = await response.json()
